making progress

+++ ft_atol.c
@@ -6,7 +6,7 @@
 /*   By: flafi <flafi@student.42.fr>                +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/08/18 22:32:43 by flafi             #+#    #+#             */
-/*   Updated: 2023/08/18 22:35:53 by flafi            ###   ########.fr       */
+/*   Updated: 2023/09/06 00:19:30 by flafi            ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -27,9 +27,9 @@ static int	ft_issign(const char *str)
 	return (0);
 }
 
-long long	ft_atol(const char *str)
+long	ft_atol(const char *str)
 {
-	long long	result;
+	long	result;
 	int			sign;
 
 	result = 0;
